[add]　サービス層の変更
