+++ flaskdb/service/ItemForm.py
@@ -1,36 +1,7 @@
-"""
-A Sample Web-DB Application for DB-DESIGN lecture
-Copyright (C) 2022 Yasuhiro Hayashi
-"""
 from flask_wtf import FlaskForm
-from wtforms import StringField, IntegerField, PasswordField, SubmitField
-from wtforms.validators import DataRequired, length
-from flaskdb.widgets import ButtonField
-
-class LoginForm(FlaskForm):
-    username = StringField(
-        "User Name",
-        validators = [
-            DataRequired(message="User Name is required."),
-            length(max=64, message="User Name should be input within 64 characters."),
-        ],
-    )
-    password = PasswordField(
-        "Password",
-        validators = [
-            DataRequired(message="Password is required."),
-        ],
-    )
-    cancel = ButtonField("Cancel")
-    submit = SubmitField("Login")
-
-    def copy_from(self, user):
-        self.username.data = user.username
-        self.password.data = user.password
-
-    def copy_to(self, user):
-        user.username = self.username.data
-        user.password = self.password.data
+from wtforms import StringField, IntegerField, SubmitField
+from wtforms.validators import DataRequired
+from flaskdb.service.widgets import ButtonField
 
 class AddItemForm(FlaskForm):
     itemname = StringField(
